# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gbif/taxa:
    get:
      operationId: searchGbifTaxa
      tags: [gbif]
      summary: Search taxa
      description: Search for taxa in the local GBIF dataset
      parameters:
        - name: q
          in: query
          description: Scientific or canonical name query
          schema:
            type: string
        - name: rank
          in: query
          description: Taxonomic rank filter
          schema:
            type: string
        - name: kingdom
          in: query
          description: Kingdom filter
          schema:
            type: string
        - name: iucnStatus
          in: query
          description: IUCN Red List status filter
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum results (default 20, max 100)
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated list of taxa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxaPaginatedResponse"
  /gbif/taxa/{taxonKey}:
    get:
      operationId: getGbifTaxon
      tags: [gbif]
      summary: Get taxon details
      description: Get full details for a specific taxon by its GBIF key
      parameters:
        - name: taxonKey
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Taxon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxon"
        "404":
          description: Taxon not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gbif/occurrences:
    get:
      operationId: searchGbifOccurrences
      tags: [gbif]
      summary: Search occurrences
      description: Search occurrence records from the local dataset
      parameters:
        - name: taxonKey
          in: query
          schema:
            type: integer
        - name: countryCode
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated list of occurrences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccurrencePaginatedResponse"
  /gbif/regions:
    get:
      operationId: listGbifRegions
      tags: [gbif]
      summary: List regions
      description: List all regions with occurrence and species counts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"
  /cognee/status:
    get:
      operationId: getCogneeStatus
      tags: [cognee]
      summary: Cognee service status
      description: Returns the status of the Cognee knowledge graph service including node/edge counts
      responses:
        "200":
          description: Cognee status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CogneeStatus"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cognee/search:
    get:
      operationId: searchCognee
      tags: [cognee]
      summary: Search Cognee knowledge graph
      description: Search for species and taxonomy insights using the Cognee graph
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (species name, kingdom, etc.)
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CogneeSearchResponse"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cognee/graph/species/{taxonKey}:
    get:
      operationId: getCogneeSpeciesGraph
      tags: [cognee]
      summary: Get species KG context
      description: Returns the knowledge graph context (nodes and edges) for a specific taxon
      parameters:
        - name: taxonKey
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Species graph context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CogneeGraphContext"
        "404":
          description: Taxon not found in graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cognee/mcp:
    get:
      operationId: getMcpInfo
      tags: [cognee]
      summary: MCP server info and tools
      description: Returns MCP server metadata and available tool definitions
      responses:
        "200":
          description: MCP server info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServerInfo"
    post:
      operationId: callMcpTool
      tags: [cognee]
      summary: Call MCP tool
      description: Execute a JSON-RPC 2.0 request against the MCP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpRequest"
      responses:
        "200":
          description: JSON-RPC response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpResponse"
...
components:
  schemas:
...
    Taxon:
      type: object
      properties:
        taxonKey:
          type: integer
        parentKey:
          type: integer
          nullable: true
        rank:
          type: string
        kingdom:
          type: string
          nullable: true
        phylum:
          type: string
          nullable: true
        class:
          type: string
          nullable: true
        order:
          type: string
          nullable: true
        family:
          type: string
          nullable: true
        genus:
          type: string
          nullable: true
        species:
          type: string
          nullable: true
        scientificName:
          type: string
        canonicalName:
          type: string
          nullable: true
        vernacularName:
          type: string
          nullable: true
        iucnStatus:
          type: string
          nullable: true
        numOccurrences:
          type: integer
        extinct:
          type: string
          nullable: true
      required:
        - taxonKey
        - rank
        - scientificName
        - numOccurrences
    TaxaPaginatedResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Taxon"
      required:
        - total
        - offset
        - limit
        - results
    Occurrence:
      type: object
      properties:
        id:
          type: integer
        taxonKey:
          type: integer
        countryCode:
          type: string
          nullable: true
        decimalLatitude:
          type: number
          nullable: true
        decimalLongitude:
          type: number
          nullable: true
        year:
          type: integer
          nullable: true
        month:
          type: integer
          nullable: true
        basisOfRecord:
          type: string
          nullable: true
      required:
        - id
        - taxonKey
    OccurrencePaginatedResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Occurrence"
      required:
        - total
        - offset
        - limit
        - results
    Region:
      type: object
      properties:
        countryCode:
          type: string
        countryName:
          type: string
        occurrenceCount:
          type: integer
        speciesCount:
          type: integer
        lastSynced:
          type: string
          nullable: true
      required:
        - countryCode
        - countryName
        - occurrenceCount
        - speciesCount
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    CogneeStatus:
      type: object
      properties:
        status:
          type: string
        nodeCount:
          type: integer
        edgeCount:
          type: integer
        nodesByType:
          type: object
          additionalProperties:
            type: integer
        taxaLoaded:
          type: integer
        loaded:
          type: boolean
        loadError:
          type: string
          nullable: true
      required:
        - status
        - nodeCount
        - edgeCount
        - taxaLoaded
        - loaded
    CogneeSearchResult:
      type: object
      properties:
        externalId:
          type: string
        nodeType:
          type: string
        label:
          type: string
        properties:
          type: object
        relatedTo:
          type: array
          items:
            type: string
        score:
          type: number
      required:
        - externalId
        - nodeType
        - label
        - relatedTo
        - score
    CogneeSearchResponse:
      type: object
      properties:
        query:
          type: string
        count:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/CogneeSearchResult"
      required:
        - query
        - count
        - results
    CogneeGraphNode:
      type: object
      properties:
        externalId:
          type: string
        nodeType:
          type: string
        label:
          type: string
        properties:
          type: object
      required:
        - externalId
        - nodeType
        - label
    CogneeGraphEdge:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        edgeType:
          type: string
        weight:
          type: number
      required:
        - from
        - to
        - edgeType
        - weight
    CogneeGraphContext:
      type: object
      properties:
        taxonKey:
          type: integer
        label:
          type: string
        nodeType:
          type: string
        properties:
          type: object
        nodeCount:
          type: integer
        edgeCount:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CogneeGraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/CogneeGraphEdge"
      required:
        - taxonKey
        - label
        - nodeType
        - nodeCount
        - edgeCount
        - nodes
        - edges
    McpTool:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        inputSchema:
          type: object
      required:
        - name
        - description
        - inputSchema
    McpServerInfo:
      type: object
      properties:
        server:
          type: object
        tools:
          type: array
          items:
            $ref: "#/components/schemas/McpTool"
        capabilities:
          type: object
      required:
        - server
        - tools
    McpRequest:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          type: string
          nullable: true
        method:
          type: string
        params:
          type: object
      required:
        - method
    McpResponse:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          type: string
          nullable: true
        result:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
      required:
        - jsonrpc