# lib/api-zod/src/generated/types/mcpServerInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { McpServerInfoCapabilities } from "./mcpServerInfoCapabilities";
import type { McpServerInfoServer } from "./mcpServerInfoServer";
import type { McpTool } from "./mcpTool";

export interface McpServerInfo {
  server: McpServerInfoServer;
  tools: McpTool[];
  capabilities?: McpServerInfoCapabilities;
}